Fixing warnings.

+++ cocos2d-ui/CCBReader/CCBReader.m
@@ -1255,7 +1255,7 @@ static inline float readFloat(CCBReader *self)
         // Class was not found. Maybe it's a Swift class?
         // See http://stackoverflow.com/questions/24030814/swift-language-nsclassfromstring
         NSString *appName = [[NSBundle mainBundle] objectForInfoDictionaryKey:@"CFBundleName"];
-        NSString *classStringName = [NSString stringWithFormat:@"_TtC%d%@%d%@", appName.length, appName, className.length, className];
+        NSString *classStringName = [NSString stringWithFormat:@"_TtC%d%@%d%@", (int)appName.length, appName, (int)className.length, className];
         class = NSClassFromString(classStringName);
     }
     if (!class)
